refactor: relocate shell components to shared module, simplify sidebar
- Move AppState.kt and NavigationSidebar.kt to shared module
- Remove duplicate imports and redundant logic from desktop main
- Reduce NavigationSidebar complexity by eliminating hard‑coded navigation paths
- Adjust imports in desktop module to reference shared implementations
- Clean up unused theme mode checks in Main.kt
- Ensure backward compatibility of existing navigation behaviour
